refactor: update e2e encryption tests
